WebDevelopment/ZMTeam/backend/pyproject.toml
[tool.poetry]
name = "zmteam"
version = "0.1.0"
description = ""
authors = ["YóUnǎi <c0d3r.nodiru.gaji@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standart"], version = "^0.115.0"}
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.5.2"
uvicorn = "^0.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
